replaced powl command, not all systems have this

+++ domoticz/main/Helper.cpp
@@ -258,7 +258,7 @@ double ConvertToFahrenheit(const double Celsius)
 
 double RoundDouble(const long double invalue, const short numberOfPrecisions)
 {
-	long long p = (long long) powl(10.0L, numberOfPrecisions);
+	long long p = (long long) pow(10.0L, numberOfPrecisions);
 	double ret= (long long)(invalue * p + 0.5L) / (double)p;
 	return ret;
 }
